implemented much

+++ helper.py
@@ -94,7 +94,7 @@ def gen_batch_function(data_folder, image_shape):
                 images.append(image)
                 gt_images.append(gt_image)
 
-            yield np.array(images), np.array(gt_images)
+            yield np.array(images).astype(float), np.array(gt_images).astype(float)
     return get_batches_fn
